Gotten ToomCook3 multi-threading. Need to add some code to inside the algorithm to set a convenient threading cut-off point (small values are entertaining as memory usage climbs into the stratosphere, but not very practical for performance).
Still a concern as I had some errors on the locking code when ToomCook3 called the BZ algorithm during early testing. The code was a total mess at the time. I've tidied things up and the error hasn't reoccurred, but as I don't know exactly what caused it in the first place I am not convinced it has gone away.